This fixes two long standing bugs which kept a mounted file system busy and unable to be unmounted.  The cause of the problem is two-fold.
First, sys_access() is broken because it doesn't release the inode it
retrieved by its call to namei().

Second, the reference count for the program's text inode
"current->t_inode" is not properly tracked by fork(), exec(), and
exit(). This left the inode marked as in use, even after a program exited.

When umount() attempted to unmount the device, it found inodes still
referencing it, so it returned EBUSY.  Another nasty effect of this
bug was that every time a unique program was executed, it permanently
seized an inode from the kernel's inode table. Because the inode was
never released, it was possible to eventually run out of inodes.

To exercise the bug, execute a program from a mounted partition, or
invoke access() on a file on that partition. Try this -- boot with an
old boot/root floppy and mount your hard drive elks partition on /mnt.
The following command will then exercise both the access() and t_inode
bugs:

          # exec /mnt/bin/sh

After exiting the shell and logging in again, you will be unable to
unmount the partition.

   < Thomas McWilliams -- tmwo@users.sourceforge.net >